+++ tests/formatter/issue_0016_force_line_break_by_using_comment.out.cmake
@@ -1,0 +1,15 @@
+add_custom_command(
+    OUTPUT src.tar.zst
+    DEPENDS ${SRC}
+    COMMAND
+        set -euo pipefail # line comment
+        && tar -c src | zstd --ultra -22 > src.tar.zst #
+)
+
+add_custom_command(
+    OUTPUT src.tar.zst
+    DEPENDS ${SRC}
+    COMMAND
+        set -euo pipefail #[[bracket comment]]
+        && tar -c src | zstd --ultra -22 > src.tar.zst #
+)