+++ tests/formatter/comments_with_quoted_strings_in_unknown_custom_commands_with_tabs.in.cmake
@@ -1,0 +1,19 @@
+### {indent: tabs}
+function(foo)
+  bar(
+    # some comment with "quoted string"
+    # some comment with "quoted string"
+    # some comment with "quoted string"
+    # some comment with "quoted string"
+  )
+endfunction()
+
+
+function(foo)
+  bar(
+    #[[some comment with "quoted string"]]
+    #[[some comment with "quoted string"]]
+    #[[some comment with "quoted string"]]
+    #[[some comment with "quoted string"]]
+  )
+endfunction()