+++ tests/test_custom_command_formatting.py
@@ -38,7 +38,7 @@ some_custom_command_without_keyworded_arguments(
 
     parsed = parser_with_postprocessing.parse(given)
     definitions = get_just_definitions(find_custom_command_definitions(parsed))
-    dumper = Dumper(width=80, custom_command_definitions=definitions)
+    dumper = Dumper(width=80, indent_size=4, custom_command_definitions=definitions)
 
     formatted = dumper.visit(parsed)
 
@@ -75,7 +75,7 @@ some_custom_command_without_keyworded_arguments(long____________________________
 
     parsed = parser_with_postprocessing.parse(given)
     definitions = get_just_definitions(find_custom_command_definitions(parsed))
-    dumper = Dumper(width=80, custom_command_definitions=definitions)
+    dumper = Dumper(width=80, indent_size=4, custom_command_definitions=definitions)
 
     formatted = dumper.visit(parsed)
 
@@ -108,7 +108,7 @@ endfunction()
 
     parsed = parser_with_postprocessing.parse(given)
     definitions = get_just_definitions(find_custom_command_definitions(parsed))
-    dumper = Dumper(width=80, custom_command_definitions=definitions)
+    dumper = Dumper(width=80, indent_size=4, custom_command_definitions=definitions)
 
     formatted = dumper.visit(parsed)